pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "aword"
version = "0.0.1"
description = "Slack bot to interact with OpenAI GPT."
readme = "README.md"
authors = [
    { name = "Joan Reyero", email="joan@crowd.dev" },
    { name = "Juan Reyero", email="juan@juanreyero.com" }
]
requires-python = ">=3.8"
# license = { file="LICENSE" }
dependencies = [
    "openai >= 0.27.2",
    "qdrant-client >= 1.1.0",
    "slack-bolt >= 1.18.0",
    "python-dotenv >= 1.0.0",
    "tqdm >= 4.65.0",
    "pytz >= 2022.7.1",
    "mistune",
    "tenacity"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
    "jedi >= 0.18.1",
    "black >= 23.3.0",
    "pylint >= 2.13.9",
    "pytest >= 7.0.0",
    "yapf >= 0.32.0",
    "build",
    "twine",
    "catprompt"
]

[project.scripts]
ingest-notion = "aword.sources.notion:ingest"
ingest-linear = "aword.sources.linear:ingest"
ingest-crowddev = "aword.sources.crowddev:ingest"
aword = "aword.chat:main"

[tool.pytest.ini_options]
addopts = "--ignore=build --ignore=doc --ignore=flymake"
pythonpath = [
  "."
]
